Make the java_maxmetaspace_size param optional.
The default behaviour on most JVM versions is to have no limit for the
Max Meta Space. While I understand wanting to set and limit it, in
v3.0.0 this change is listed as an enhancement while actually being a
breaking change.

When one only reads the changelog without looking at the code,
applications using more then the low default of 128MB will crash.